separate policy resolver
